frictionless field

Based on the class input vector x, the appropriate fr_*() function will be called to create a frictionless field descriptor object (fr_field).

Usage

fr_field(x, name, ...)

fr_string(x, name, ...)

Arguments

x

a character, factor, numeric, integer, logical, or Date vector

name

string required name metadata descriptor as a string

...

<dynamic-dots> other optional metadata descriptors (title, description, example)

Value

a fr_field vector

Details

Only classes listed in this table (and in this order) are used to automatically select the correct fr_*() function.

R class	fr type	fr_*()

character, factor^	string	fr_string()

numeric, integer	number	fr_number()

logical	boolean	fr_boolean()

Date	date	fr_date()

^Levels of factor columns are also captured in the "enum" item of the "constraints" attribute list.

To convert a class not specifically listed to a frictionless type or to parse character vectors for a specific frictionless type, use one of the fr_*() functions instead.

Examples

fr_field(letters, name = "example_string")
#> <fr_field[26]>
#> example_string (fr_string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z

# TODO fr_field(factor(LETTERS), name = "example_factor")

fr_field(1:26, name = "example_numbers")
#> <fr_field[26]>
#> example_numbers (fr_number)
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26

fr_field(c(TRUE, FALSE, TRUE), name = "example_logical")
#> <fr_field[3]>
#> example_logical (fr_boolean)
#> [1]  TRUE FALSE  TRUE

fr_field(as.Date(c("2013-08-15", "1986-04-29", "1986-06-10")), name = "example_date")
#> <fr_field[3]>
#> example_date (fr_date)
#> [1] 2013-08-15 1986-04-29 1986-06-10

# a more realistic field with metadata
uid <-
  fr_field(
    replicate(34, paste0(sample(c(letters,0:9), 8, TRUE), collapse = "")),
    name = "uid",
    title = "Unique Identifier",
    description = "Consists of 8 random characters from (a-z) and (0-9)"
)

# examine metadata in `fr_field` objects using base R functions:

# the default print method includes the name and frictionless type,
# but otherwise prints the same as the underlying R type (here, a character vector):
uid
#> <fr_field[34]>
#> uid (fr_string)
#>  [1] idl9tt1a qw3uktgm k60errij dtqjl26g 8jviqewr 0e3teeav 577wgbfy 247j8mms
#>  [9] 3467s1sx qj3qgzuv rbc8hlqo pge0fkgg 4uiv8nvf lhdjhtmn 42jokfhm krv678g3
#> [17] qpz6w01x sa1rcp4a d2fdegz8 gbhrl5wt kmj0vkdg dlsz8njg d6a6rgth pnqy1t5u
#> [25] 5xyxpz23 q7n4rv5a ie8b819b 5gxik24c 26ka2ls6 gdxu1dr9 a73k11m4 zi7pnqgh
#> [33] etkd0cao mo7eyppp

str(uid)
#> uid (fr_string)
#>  fr_field [1:34] idl9tt1a, qw3uktgm, k60errij, dtqjl26g, 8jviqewr, 0e3teeav...
#>  @ name       : chr "uid"
#>  @ title      : chr "Unique Identifier"
#>  @ description: chr "Consists of 8 random characters from (a-z) and (0-9)"
#>  @ type       : chr "string"

attributes(uid)
#> $name
#> [1] "uid"
#> 
#> $title
#> [1] "Unique Identifier"
#> 
#> $description
#> [1] "Consists of 8 random characters from (a-z) and (0-9)"
#> 
#> $type
#> [1] "string"
#> 
#> $class
#> [1] "fr_field"   "vctrs_vctr"
#> 

# retrieve metadata descriptors with `attr()`:
attr(uid, "title")
#> [1] "Unique Identifier"

# using glue_data makes it easy to write clear documentation and messages using metadata fields:
glue::glue_data(attributes(uid), "`{name}` (a.k.a. '{title}') is a {type} field. {description}.")
#> `uid` (a.k.a. 'Unique Identifier') is a string field. Consists of 8 random characters from (a-z) and (0-9).
fr_string(letters, name = "letters")
#> <fr_field[26]>
#> letters (fr_string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z
fr_string(letters, name = "letters", title = "Letters")
#> <fr_field[26]>
#> letters (fr_string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z