Coerce character, factor, numeric, logical, and Date vectors into fr_field objects

The supported classes of R objects are converted to the corresponding frictionless type:

R class	fr type

character(), factor()	string*

numeric(), integer()	number

logical()	boolean

Date	date

Usage

as_fr_field(x, ...)

Arguments

x

a character, factor, numeric, integer, logical, or Date vector

...

name, title, or description property (name is required)

Value

a fr_field object

Details

*If a factor(), the enum constraint is set to the levels of the factor in R.

Examples

as_fr_field(1:10, "example_integer") # -> frictionless numeric
#> example_integer (numeric)
#>  [1]  1  2  3  4  5  6  7  8  9 10
as_fr_field((1:10) * 0.1, "example_double") # -> frictionless numeric
#> example_double (numeric)
#>  [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0
as_fr_field(letters, "example_character") # -> frictionless string
#> example_character (string)
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
#> [20] "t" "u" "v" "w" "x" "y" "z"
as_fr_field(factor(letters), "example_factor") # -> frictionless string with enum constraints
#> example_factor (string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z
#> Levels: a b c d e f g h i j k l m n o p q r s t u v w x y z
as_fr_field(c(TRUE, FALSE, TRUE), "example_logical") # -> frictionless boolean
#> example_logical (boolean)
#> [1]  TRUE FALSE  TRUE
as_fr_field(as.Date(c("2023-04-23", "2004-12-31")), "example_date") # -> frictionless date
#> example_date (date)
#> [1] "2023-04-23" "2004-12-31"

# an example `fr_field` object
uid <-
  replicate(34, paste0(sample(c(letters, 0:9), 8, TRUE), collapse = "")) |>
  as_fr_field(
    name = "uid",
    title = "Unique Identifier",
  )

# set properties of the `fr_field`
S7::prop(uid, "description")
#> character(0)
uid@description <- "Consists of 8 random characters from (a-z) and (0-9)"

# inspect properties of the `fr_field`
uid
#> uid (string)
#>  [1] "ek5qb386" "t3p23zaf" "1a189be3" "ob96v29g" "802rpz3w" "bcqgo6bp"
#>  [7] "jjw5uwo4" "s3j5mc9c" "fjr2qtiz" "67l40tb7" "lecnf9es" "hm2p2c7j"
#> [13] "z7wrkfue" "oimmfh4m" "b8frr8ti" "8eh2za12" "9pvanc3r" "j18kfh7f"
#> [19] "6p5yi32r" "ejl2plj6" "ppgl2psv" "c1uw2i68" "sn7en4fo" "ti22rq0o"
#> [25] "4h6gqekk" "d2af190e" "4u0an1zc" "cg77tjf7" "1buu01z3" "x0w3l9nk"
#> [31] "efo01upg" "47tu61ye" "llnzcmqh" "eelbaejk"
str(uid)
#> <fr::fr_field>
#>  @ value      : chr [1:34] "ek5qb386" "t3p23zaf" "1a189be3" "ob96v29g" "802rpz3w" ...
#>  @ name       : chr "uid"
#>  @ type       : chr "string"
#>  @ title      : chr "Unique Identifier"
#>  @ description: chr "Consists of 8 random characters from (a-z) and (0-9)"
#>  @ constraints: list()