frictionless field descriptor

Automatically create a frictionless field descriptor object (fr_field) with the appropriate frictionless type, format, and constraints based on the class of the input vector x:

Usage

fr_field(x, name = character(), ...)

is_fr_field(x)

Arguments

x

For fr_field, a character, factor, numeric, integer, logical, or Date vector

For is_fr_field, an object to test

name

metadata name descriptor as a string; if NULL, it will be set automatically using rlang::as_label(), which is usually nondescript

...

<dynamic-dots> other user-set optional metadata descriptors (e.g., title, description, example)

Value

a fr_field vector

Details

R class	fr type	format	constraints

character(), factor()	string

numeric(), integer()	number

logical()	boolean

Date	date

Only these specific classes are automatically coerced. To convert a class not specifically listed to a frictionless type or to parse character vectors for a specific frictionless type, use one of the fr_field_*() functions instead.

Examples

fr_field(letters, name = "example_string")
#> <fr_field[26]>
#> example_string (fr_string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z

fr_field(factor(LETTERS), name = "example_factor")
#> <fr_field[26]>
#> example_factor (fr_string)*
#>  [1] A B C D E F G H I J K L M N O P Q R S T U V W X Y Z

fr_field(1:26, name = "example_numbers")
#> <fr_field[26]>
#> example_numbers (fr_number)
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26

fr_field(c(TRUE, FALSE, TRUE), name = "example_logical")
#> <fr_field[3]>
#> example_logical (fr_boolean)
#> [1]  TRUE FALSE  TRUE

fr_field(as.Date(c("2013-08-15", "1986-04-29", "1986-06-10")), name = "example_date")
#> <fr_field[3]>
#> example_date (fr_date)
#> [1] 2013-08-15 1986-04-29 1986-06-10

# a more realistic field with metadata
uid <-
  fr_field(
    replicate(34, paste0(sample(c(letters,0:9), 8, TRUE), collapse = "")),
    name = "uid",
    title = "Unique Identifier",
    description = "Consists of 8 random characters from (a-z) and (0-9)"
)

# examine metadata in `fr_field` objects using base R functions:

# the default print method includes the name and frictionless type,
# but otherwise prints the same as the underlying R type (here, a character vector):
uid
#> <fr_field[34]>
#> uid (fr_string)
#>  [1] evbnar6o zubux8uv ozn4gkwu ewb906tg th0os796 x41kbx3l hyuy3kix i6juctc1
#>  [9] 27yxlhub 6xqwhzov 5lbkjt2o fu8en5l8 8dt8hpgf ggh2yaf1 e93f4x12 eeadxp4d
#> [17] 8s753wbb zapcas1o 226b44j5 agn9v6hj 3gl3hvxh wk82lks9 bpojiy00 zd9sqjn8
#> [25] vx7o0c6u 87gucypk pipfd2bf xjon1g0k w1w6xtgs yq6nmpo0 xwba9b07 p7awupmr
#> [33] v670ot4j naxxrfzf

str(uid)
#> uid (fr_string)
#>  field [1:34] evbnar6o, zubux8uv, ozn4gkwu, ewb906tg, th0os796, x41kbx3l, h...
#>  @ name       : chr "uid"
#>  @ type       : chr "string"
#>  @ title      : chr "Unique Identifier"
#>  @ description: chr "Consists of 8 random characters from (a-z) and (0-9)"

attributes(uid)
#> $name
#> [1] "uid"
#> 
#> $type
#> [1] "string"
#> 
#> $title
#> [1] "Unique Identifier"
#> 
#> $description
#> [1] "Consists of 8 random characters from (a-z) and (0-9)"
#> 
#> $class
#> [1] "fr_field"   "vctrs_vctr"
#> 

# retrieve metadata descriptors with `attr()`:
attr(uid, "title")
#> [1] "Unique Identifier"

# using glue_data makes it easy to write clear documentation and messages using metadata fields:
glue::glue_data(attributes(uid), "`{name}` (a.k.a. '{title}') is a {type} field. {description}.")
#> `uid` (a.k.a. 'Unique Identifier') is a string field. Consists of 8 random characters from (a-z) and (0-9).