frictionless field

Automatically create a frictionless field descriptor object (fr_field) with the appropriate frictionless type, format, and constraints based on the class of the input vector x:

Usage

fr_field(x, name, ...)

Arguments

x

a character, factor, numeric, integer, logical, or Date vector

name

required name metadata descriptor as a string

...

<dynamic-dots> other optional metadata descriptors (title, description, example)

Value

a fr_field vector

Details

| R class | fr type | format | constraints | |:------------------:|:-----------:|:--------------:| | | character, factor^ | string | | | | numeric, integer | number | | | | logical | boolean | | | | Date | date | | |

Only these specific classes are automatically coerced. To convert a class not specifically listed to a frictionless type or to parse character vectors for a specific frictionless type, use one of the fr_field_*() functions instead.

Examples

fr_field(letters, name = "example_string")
#> <fr_field[26]>
#> example_string (fr_string)
#>  [1] a b c d e f g h i j k l m n o p q r s t u v w x y z

# TODO fr_field(factor(LETTERS), name = "example_factor")

fr_field(1:26, name = "example_numbers")
#> <fr_field[26]>
#> example_numbers (fr_number)
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26

fr_field(c(TRUE, FALSE, TRUE), name = "example_logical")
#> <fr_field[3]>
#> example_logical (fr_boolean)
#> [1]  TRUE FALSE  TRUE

fr_field(as.Date(c("2013-08-15", "1986-04-29", "1986-06-10")), name = "example_date")
#> <fr_field[3]>
#> example_date (fr_date)
#> [1] 2013-08-15 1986-04-29 1986-06-10

# a more realistic field with metadata
uid <-
  fr_field(
    replicate(34, paste0(sample(c(letters,0:9), 8, TRUE), collapse = "")),
    name = "uid",
    title = "Unique Identifier",
    description = "Consists of 8 random characters from (a-z) and (0-9)"
)

# examine metadata in `fr_field` objects using base R functions:

# the default print method includes the name and frictionless type,
# but otherwise prints the same as the underlying R type (here, a character vector):
uid
#> <fr_field[34]>
#> uid (fr_string)
#>  [1] 6dwejxq1 d3yatir3 ehi5z3f4 x9htloyo hf961xy1 dsbejrdp 2thuzdbd 06xudf0l
#>  [9] bram0cam fkkze74r qrbxdawb 6j3blk26 dvnakg8y ph3aceyy 0jtvnrzm h68llhxs
#> [17] gjawhg77 hrn1yqku za275fgs ao6padnt h24d4x03 hkis103w egw1sivh 6bw94kbt
#> [25] j6muplrr wo60d5lw uq6d5zzk cr4j3xu5 56fmor9y ge8oukoq 9qtx2byu ct23e99c
#> [33] omalnhpo tqenddlh

str(uid)
#> uid (fr_string)
#>  fr_field [1:34] 6dwejxq1, d3yatir3, ehi5z3f4, x9htloyo, hf961xy1, dsbejrdp...
#>  @ name       : chr "uid"
#>  @ title      : chr "Unique Identifier"
#>  @ description: chr "Consists of 8 random characters from (a-z) and (0-9)"
#>  @ type       : chr "string"

attributes(uid)
#> $name
#> [1] "uid"
#> 
#> $title
#> [1] "Unique Identifier"
#> 
#> $description
#> [1] "Consists of 8 random characters from (a-z) and (0-9)"
#> 
#> $type
#> [1] "string"
#> 
#> $class
#> [1] "fr_field"   "vctrs_vctr"
#> 

# retrieve metadata descriptors with `attr()`:
attr(uid, "title")
#> [1] "Unique Identifier"

# using glue_data makes it easy to write clear documentation and messages using metadata fields:
glue::glue_data(attributes(uid), "`{name}` (a.k.a. '{title}') is a {type} field. {description}.")
#> `uid` (a.k.a. 'Unique Identifier') is a string field. Consists of 8 random characters from (a-z) and (0-9).